#18515: Moved conv and pool tests to new folders, changed codeowners
