+++ tests/ttnn/nightly/unit_tests/operations/conv/test_conv3d.py
@@ -9,7 +9,7 @@ import ttnn
 import torch.nn as nn
 from tests.ttnn.utils_for_testing import check_with_pcc
 
-from tests.ttnn.unit_tests.operations.test_conv3d import (
+from tests.ttnn.unit_tests.operations.conv.test_conv3d import (
     setup_conv3d_test,
     create_conv3d_config,
     prepare_weights,
