+++ tests/ttnn/unit_tests/operations/conv/test_prepare_conv_weights.py
