+++ tests/ttnn/unit_tests/operations/pool/test_downsample.py
