+++ tests/ttnn/unit_tests/operations/pool/test_maxpool2d.py
